+++ src/node/attr.cpp
@@ -171,7 +171,7 @@ auto get_attr(tinytc_attr_t dict, std::string_view name) -> tinytc_attr_t {
 extern "C" {
 
 tinytc_status_t tinytc_array_attr_get(tinytc_attr_t *attr, tinytc_compiler_context_t ctx,
-                                      uint32_t array_size, const tinytc_attr_t *array) {
+                                      size_t array_size, const tinytc_attr_t *array) {
     if (attr == nullptr || ctx == nullptr || (array_size != 0 && array == nullptr)) {
         return tinytc_status_invalid_arguments;
     }
@@ -188,14 +188,14 @@ tinytc_status_t tinytc_boolean_attr_get(tinytc_attr_t *attr, tinytc_compiler_con
 }
 
 tinytc_status_t tinytc_dictionary_attr_get(tinytc_attr_t *attr, tinytc_compiler_context_t ctx,
-                                           uint32_t items_size, tinytc_named_attr_t *items) {
+                                           size_t items_size, tinytc_named_attr_t *items) {
     TINYTC_CHECK_STATUS(tinytc_dictionary_attr_sort(items_size, items));
     return tinytc_dictionary_attr_get_with_sorted(attr, ctx, items_size, items);
 }
 
 tinytc_status_t tinytc_dictionary_attr_get_with_sorted(tinytc_attr_t *attr,
                                                        tinytc_compiler_context_t ctx,
-                                                       uint32_t items_size,
+                                                       size_t items_size,
                                                        const tinytc_named_attr_t *items) {
     if (attr == nullptr || ctx == nullptr || (items_size != 0 && items == nullptr)) {
         return tinytc_status_invalid_arguments;
@@ -205,7 +205,7 @@ tinytc_status_t tinytc_dictionary_attr_get_with_sorted(tinytc_attr_t *attr,
     });
 }
 
-tinytc_status_t tinytc_dictionary_attr_sort(uint32_t items_size, tinytc_named_attr_t *items) {
+tinytc_status_t tinytc_dictionary_attr_sort(size_t items_size, tinytc_named_attr_t *items) {
     return exception_to_status_code(
         [&] { dictionary_attr::sort(mutable_array_view<tinytc_named_attr_t>(items, items_size)); });
 }
@@ -219,7 +219,7 @@ tinytc_status_t tinytc_integer_attr_get(tinytc_attr_t *attr, tinytc_compiler_con
 }
 
 tinytc_status_t tinytc_string_attr_get(tinytc_attr_t *attr, tinytc_compiler_context_t ctx,
-                                       uint32_t str_length, char const *str) {
+                                       size_t str_length, char const *str) {
     if (attr == nullptr || ctx == nullptr) {
         return tinytc_status_invalid_arguments;
     }
