+++ src/node/region.hpp
@@ -66,12 +66,12 @@ struct alignas(8) tinytc_region final {
     inline auto params() { return tinytc::iterator_range_wrapper{param_begin(), param_end()}; }
     inline auto param_begin() const { return params_.begin(); }
     inline auto param_end() const { return params_.end(); }
-    inline auto param(std::int64_t pos) -> tinytc_value & { return params_[pos]; }
-    inline auto param(std::int64_t pos) const -> tinytc_value const & { return params_[pos]; }
+    inline auto param(std::size_t pos) -> tinytc_value & { return params_[pos]; }
+    inline auto param(std::size_t pos) const -> tinytc_value const & { return params_[pos]; }
     inline auto params() const {
         return tinytc::iterator_range_wrapper{param_begin(), param_end()};
     }
-    inline auto num_params() const noexcept -> std::int64_t { return params_.size(); }
+    inline auto num_params() const noexcept -> std::size_t { return params_.size(); }
     void set_params(tinytc::array_view<tinytc_data_type_t> param_types);
     void set_num_params(std::size_t num_params);
     void set_param(std::size_t idx, tinytc_data_type_t param_type);
